wip: only build versions that dont exist on dist
This PR will rework things to only build binaries that aren't already published on dist.ipfs.io.

The releases dir will be patched into the exiting CID for dist.ipfs.io to create the new site.

I assert that this will work and be rad because content-addressing.

TODO:
- [x] only build what changed
- [ ] fix website generation to not depend on a complete local releases dir
- [ ] fix publishing to patch new release into existing dist.ipfs.io CID

License: MIT
Signed-off-by: Oli Evans <oli@tableflip.io>